# perf: replace samples map with a slice