--- a/internal/agentcfg/reporter_test.go
+++ b/internal/agentcfg/reporter_test.go
@@ -77,8 +77,8 @@ func TestReportFetch(t *testing.T) {
 			Labels:    model.Labels{"etag": {Value: "abc123"}},
 			Metricset: &model.Metricset{
 				Name: "agent_config",
-				Samples: map[string]model.MetricsetSample{
-					"agent_config_applied": {Value: 1},
+				Samples: []model.MetricsetSample{
+					{Name: "agent_config_applied", Value: 1},
 				},
 			},
 		},
@@ -87,8 +87,8 @@ func TestReportFetch(t *testing.T) {
 			Labels:    model.Labels{"etag": {Value: "def456"}},
 			Metricset: &model.Metricset{
 				Name: "agent_config",
-				Samples: map[string]model.MetricsetSample{
-					"agent_config_applied": {Value: 1},
+				Samples: []model.MetricsetSample{
+					{Name: "agent_config_applied", Value: 1},
 				},
 			},
 		},