--- a/internal/model/modeldecoder/v2/decoder.go
+++ b/internal/model/modeldecoder/v2/decoder.go
@@ -628,7 +628,7 @@ func mapToMetricsetModel(from *metricset, event *model.APMEvent) bool {
 	}
 
 	if len(from.Samples) > 0 {
-		samples := make(map[string]model.MetricsetSample, len(from.Samples))
+		samples := make([]model.MetricsetSample, 0, len(from.Samples))
 		for name, sample := range from.Samples {
 			var counts []int64
 			var values []float64
@@ -640,15 +640,16 @@ func mapToMetricsetModel(from *metricset, event *model.APMEvent) bool {
 				counts = make([]int64, n)
 				copy(counts, sample.Counts)
 			}
-			samples[name] = model.MetricsetSample{
+			samples = append(samples, model.MetricsetSample{
 				Type:  model.MetricType(sample.Type.Val),
+				Name:  name,
 				Unit:  sample.Unit.Val,
 				Value: sample.Value.Val,
 				Histogram: model.Histogram{
 					Values: values,
 					Counts: counts,
 				},
-			}
+			})
 		}
 		event.Metricset.Samples = samples
 	}