<!-- Thanks for sending a pull request!

If this is your first contribution, please review and sign our contributor agreement -
https://www.elastic.co/contributor-agreement.

Guidelines:
 - Prefer small PRs, and split changes into multiple logical commits where they must
   be delivered in a single PR.
 - If the PR is incomplete and not yet ready for review, open it as a Draft.
 - Once the PR is marked ready for review it is expected to pass all tests and linting,
   and you should not force-push any changes.

See also https://github.com/elastic/apm-server/blob/main/CONTRIBUTING.md for more tips on contributing.
-->

## Motivation/summary

Reduce allocations by replacing the samples map with a slice
with perfect size.

Benchmark on heavy.ndjson:

```
name                       old time/op    new time/op    delta
ReadBatch/heavy.ndjson-16    11.2ms ± 9%    10.5ms ± 8%   -5.85%  (p=0.029 n=10+10)

name                       old speed      new speed      delta
ReadBatch/heavy.ndjson-16  35.7MB/s ± 9%  37.9MB/s ± 8%   +6.20%  (p=0.029 n=10+10)

name                       old alloc/op   new alloc/op   delta
ReadBatch/heavy.ndjson-16    1.28MB ± 0%    0.83MB ± 0%  -35.09%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
ReadBatch/heavy.ndjson-16     23.1k ± 0%     22.6k ± 0%   -2.09%  (p=0.000 n=10+9)
```

## Checklist

<!--
Delete irrelevant items. The changelog should only be updated for user-facing changes.
Once the PR is ready for review there should be no unticked boxes.
-->

- [ ] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)
- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)
- [ ] Documentation has been updated

For functional changes, consider:
- Is it observable through the addition of either **logging** or **metrics**?
- Is its use being published in **telemetry** to enable product improvement?
- Have system tests been added to avoid regression?

## How to test these changes

<!--
Explain how this PR can be tested by the reviewer: commands, dependencies, steps, etc.
If it is self-explanatory, delete this section.
-->

## Related issues

<!--
Reference the related issue(s), and make use of magic keywords where it makes sense
https://help.github.com/articles/closing-issues-using-keywords/.
-->